+++ src/features/Keybinds.cpp
@@ -5,8 +5,8 @@
 
 #include <geode.custom-keybinds/include/Keybinds.hpp>
 #include <Geode/modify/EditorUI.hpp>
-#include <utils/EditCommandExt.hpp>
 #include <Geode/binding/EditButtonBar.hpp>
+#include <utils/Editor.hpp>
 #include <utils/HolyUB.hpp>
 #include "GridScaling.hpp"