+++ src/features/MoveMenu.cpp
@@ -3,8 +3,7 @@
 #include <Geode/binding/CCMenuItemSpriteExtra.hpp>
 #include <Geode/binding/GameManager.hpp>
 #include <Geode/utils/cocos.hpp>
-#include <utils/EditCommandExt.hpp>
-#include <utils/HandleUIHide.hpp>
+#include <utils/Editor.hpp>
 #include <utils/Pro.hpp>
 
 using namespace geode::prelude;