+++ src/utils/EditorExitEvent.hpp
@@ -1,7 +1,0 @@
-#pragma once
-
-#include <Geode/loader/Event.hpp>
-
-using namespace geode::prelude;
-
-struct EditorExitEvent : public Event {};