+++ src/utils/EditorViewOnlyMode.hpp
@@ -1,8 +1,0 @@
-#pragma once
-
-#include <Geode/binding/LevelEditorLayer.hpp>
-
-using namespace geode::prelude;
-
-LevelEditorLayer* createViewOnlyEditor(GJGameLevel* level, utils::MiniFunction<CCLayer*()> returnTo);
-bool isViewOnlyEditor(LevelEditorLayer* lel);